component切り出し
